Move FrameDirectoryBuffer to the separate namespace
